+++ apps/core-app/uno.config.ts
@@ -3,10 +3,7 @@ import ri from '@iconify-json/ri/icons.json'
 import simpleIcons from '@iconify-json/simple-icons/icons.json'
 import type { IconifyJSON } from '@iconify/types'
 import presetIcons from '@unocss/preset-icons'
-// import type { Theme } from 'unocss/preset-uno'
-// import presetTheme from 'unocss-preset-theme'
 import transformerAttributifyJsx from '@unocss/transformer-attributify-jsx'
-// uno.config.ts
 import { defineConfig } from 'unocss'
 import { presetAttributify, presetUno } from 'unocss'
 
@@ -22,7 +19,7 @@ export default defineConfig({
   theme: {
     colors: {
       brand: {
-        primary: '#0c23ff'
+        primary: '#409eff'
       }
     }
   },
@@ -40,16 +37,6 @@ export default defineConfig({
         carbon: carbonIcons as IconifyJSON
       }
     })
-    // presetTheme<Theme>({
-    //   theme: {
-    //     dark: {
-    //
-    //     },
-    //     compact: {
-    //
-    //     }
-    //   }
-    // }),
   ],
   transformers: [transformerAttributifyJsx()]
 })